+++ include/boost/beast/core.hpp
@@ -12,7 +12,7 @@
 
 #include <boost/beast/core/detail/config.hpp>
 
-#include <boost/beast/core/basic_timeout_stream.hpp>
+#include <boost/beast/core/basic_stream.hpp>
 #include <boost/beast/core/bind_handler.hpp>
 #include <boost/beast/core/buffer_size.hpp>
 #include <boost/beast/core/buffer_traits.hpp>
@@ -41,10 +41,9 @@
 #include <boost/beast/core/span.hpp>
 #include <boost/beast/core/static_buffer.hpp>
 #include <boost/beast/core/static_string.hpp>
-#include <boost/beast/core/stranded_socket.hpp>
 #include <boost/beast/core/stream_traits.hpp>
 #include <boost/beast/core/string.hpp>
 #include <boost/beast/core/string_param.hpp>
-#include <boost/beast/core/timeout_stream.hpp>
+#include <boost/beast/core/tcp_stream.hpp>
 
 #endif
