+++ include/boost/beast/core/tcp_stream.hpp
@@ -7,22 +7,25 @@
 // Official repository: https://github.com/boostorg/beast
 //
 
-#ifndef BOOST_BEAST_CORE_TIMEOUT_STREAM_HPP
-#define BOOST_BEAST_CORE_TIMEOUT_STREAM_HPP
+#ifndef BOOST_BEAST_CORE_TCP_STREAM_HPP
+#define BOOST_BEAST_CORE_TCP_STREAM_HPP
 
 #include <boost/beast/core/detail/config.hpp>
-#include <boost/beast/core/basic_timeout_stream.hpp>
-#include <boost/beast/core/error.hpp>
+#include <boost/beast/core/basic_stream.hpp>
 #include <boost/asio/ip/tcp.hpp>
 
 namespace boost {
 namespace beast {
 
-/** A TCP/IP stream socket which supports timeouts and rate limits
+/** A TCP/IP stream socket with timeouts, rate limits, and executor.
+
+    @tparam Executor The type of executor to use for all completion
+    handlers which do not already have an associated executor.
+
+    @see basic_stream
 */
-using timeout_stream = basic_timeout_stream<
-    net::ip::tcp,
-    net::io_context::executor_type>;
+template<class Executor>
+using tcp_stream = basic_stream<net::ip::tcp, Executor>;
 
 } // beast
 } // boost
