+++ test/beast/core/Jamfile
@@ -18,7 +18,7 @@ local SOURCES =
     _detail_variant.cpp
     _detail_varint.cpp
     async_op_base.cpp
-    basic_timeout_stream.cpp
+    basic_stream.cpp
     bind_handler.cpp
     buffer_size.cpp
     buffer_traits.cpp
@@ -49,11 +49,10 @@ local SOURCES =
     span.cpp
     static_buffer.cpp
     static_string.cpp
-    stranded_socket.cpp
     stream_traits.cpp
     string.cpp
     string_param.cpp
-    timeout_stream.cpp
+    tcp_stream.cpp
     ;
 
 local RUN_TESTS ;
